refactor: Replace legacy SWR hooks with centralized service.ts hooks
- Remove hooks/main/queries folder with custom SWR hooks
- Replace useGoalList with useGoals from service.ts
- Replace useGoalWithSubGoalTodo with useGoalWithSubGoals from service.ts
- Replace useTodoList with useSubGoalTodos from service.ts
- Migrate data conversion logic inline to components
- Maintain backward compatibility with existing component interfaces
- Consolidate all API calls through centralized service layer
- Remove duplicate SWR logic and improve maintainability
