+++ components/main/GoalCard/GoalCard.tsx
@@ -2,15 +2,19 @@
 
 import useGoalStore from "@/stores/useGoalStore";
 import TodoList from "../TodoList/TodoList";
-import useGoalWithSubGoalTodo from "@/hooks/main/queries/useGoalWithSubGoalTodo";
-import { GoalWithSubGoalTodoRs } from "@/api/generated/motimo/Api";
+import { useGoalWithSubGoals, useSubGoalTodos } from "@/service";
+import {
+  GoalWithSubGoalTodoRs,
+  TodoRsStatusEnum,
+} from "@/api/generated/motimo/Api";
 import GoalTitleArea from "../GoalTitleArea/GoalTitleArea";
 import GoalInfo from "@/components/shared/GoalInfo/GoalInfo";
 import TodoBottomSheet, {
   TodoBottomSheetProps,
   TodoInfoForSubmission,
 } from "@/components/shared/BottomSheets/TodoBottomSheet/TodoBottomSheet";
-import useTodoList from "@/hooks/main/queries/useTodoList";
+import { TodoListProps } from "../TodoList/TodoList";
+import { TodoItemsInfo } from "@/types/todoList";
 import { useEffect, useRef, useState } from "react";
 import { createNewGoal } from "@/lib/main/goalFetching";
 import { createNewTodoOnSubGoal } from "@/lib/main/subGoalFetching";
@@ -19,17 +23,49 @@ import useActiveTodoBottomSheet from "@/stores/useActiveTodoBottomSheet";
 import useModal from "@/hooks/useModal";
 import { date2StringWithSpliter } from "@/utils/date2String";
 
+// Define the converted data type
+type ConvertedGoalWithSubGoalTodo = Omit<GoalWithSubGoalTodoRs, "subGoals"> & {
+  subGoals: TodoListProps[];
+};
+
 interface GoalCardProps {
   initSubGoalTodo?: GoalWithSubGoalTodoRs;
 }
 
 const GoalCard = ({ initSubGoalTodo }: GoalCardProps) => {
   const { goalId } = useGoalStore();
-  const { data: goalWithSubGoalTodo } = useGoalWithSubGoalTodo(goalId || "", {
+  const { data: rawGoalData } = useGoalWithSubGoals(goalId || "", {
     fallbackData: initSubGoalTodo,
   });
+
+  // Convert raw goal data to the format expected by the component
+  const goalWithSubGoalTodo: ConvertedGoalWithSubGoalTodo = {
+    ...rawGoalData,
+    subGoals:
+      rawGoalData?.subGoals?.map((subGoalInfo) => {
+        const todosInSubGoal: TodoItemsInfo[] =
+          subGoalInfo.todos?.map((todoInfo) => ({
+            id: todoInfo.id ?? "",
+            title: todoInfo.title ?? "",
+            checked: todoInfo.status === TodoRsStatusEnum.COMPLETE,
+            reported: todoInfo.todoResultId ? true : false,
+            targetDate: todoInfo.date ? new Date(todoInfo.date) : new Date(),
+          })) ?? [];
+
+        return {
+          subGoalId: subGoalInfo.id ?? "",
+          initTodoTotalLen: todosInSubGoal.length,
+          initTodoItemsInfo: todosInSubGoal,
+          subGoal: subGoalInfo.title,
+          initTodoCheckedLen: todosInSubGoal.filter(
+            (todoInfo) => todoInfo.checked,
+          ).length,
+        };
+      }) ?? [],
+  } as ConvertedGoalWithSubGoalTodo;
+
   const [newTodo, setNewTodo] = useState<TodoInfoForSubmission | null>(null);
-  const { mutate } = useTodoList(newTodo?.subGoalId ?? "");
+  const { mutate } = useSubGoalTodos(newTodo?.subGoalId ?? "");
   const { isActive, setIsActive, initContent } = useActiveTodoBottomSheet();
   const { isOpened: isModalOpened } = useModal();
